SFTP: more types to uint32_t
The 'num_names' field in the SSH_FXP_NAME response is an unsigned 32bit
value so we make sure to treat it like that.

+++ src/sftp.h
@@ -112,7 +112,7 @@ struct _LIBSSH2_SFTP_HANDLE
         } file;
         struct _libssh2_sftp_handle_dir_data
         {
-            unsigned long names_left;
+            uint32_t names_left;
             void *names_packet;
             char *next_name;
         } dir;
